refactor: request dsl to separate directory
