+++ src/main/java/ru/tinkoff/gatling/javaapi/request/PublishDslBuilder.java
@@ -1,88 +1,85 @@
-package ru.tinkoff.gatling.javaapi;
+package ru.tinkoff.gatling.javaapi.request;
+
+import java.util.Map;
+import static io.gatling.javaapi.core.internal.Expressions.*;
 
 import io.gatling.javaapi.core.ActionBuilder;
-import ru.tinkoff.gatling.amqp.request.AmqpProtocolMessage;
 import scala.Tuple2;
 
-import java.util.Arrays;
-import java.util.Map;
+public class PublishDslBuilder implements ActionBuilder {
 
-import static io.gatling.javaapi.core.internal.Expressions.*;
-import static io.gatling.javaapi.core.internal.Expressions.toStringExpression;
+    private ru.tinkoff.gatling.amqp.request.PublishDslBuilder wrapped;
 
-public class RequestReplyDslBuilder implements ActionBuilder {
-    private ru.tinkoff.gatling.amqp.request.RequestReplyDslBuilder wrapped;
-
-    public RequestReplyDslBuilder(ru.tinkoff.gatling.amqp.request.RequestReplyDslBuilder wrapped) {
+    public PublishDslBuilder(ru.tinkoff.gatling.amqp.request.PublishDslBuilder wrapped) {
         this.wrapped = wrapped;
     }
 
-    public RequestReplyDslBuilder messageId(String value) {
+    public PublishDslBuilder messageId(String value) {
         this.wrapped = wrapped.messageId(toStringExpression(value));
         return this;
     }
 
-    public RequestReplyDslBuilder priority(Integer value) {
+    public PublishDslBuilder priority(Integer value) {
         this.wrapped = wrapped.priority(toIntExpression(value.toString()));
         return this;
     }
 
-    public RequestReplyDslBuilder contentType(String value) {
+    public PublishDslBuilder contentType(String value) {
         this.wrapped = wrapped.contentType(toStringExpression(value));
         return this;
     }
 
-    public RequestReplyDslBuilder contentEncoding(String value) {
+    public PublishDslBuilder contentEncoding(String value) {
         this.wrapped = wrapped.contentEncoding(toStringExpression(value));
         return this;
     }
 
-    public RequestReplyDslBuilder correlationId(String value) {
+    public PublishDslBuilder correlationId(String value) {
         this.wrapped = wrapped.correlationId(toStringExpression(value));
         return this;
     }
 
-    public RequestReplyDslBuilder replyTo(String value) {
+    public PublishDslBuilder replyTo(String value) {
         this.wrapped = wrapped.replyTo(toStringExpression(value));
         return this;
     }
 
-    public RequestReplyDslBuilder expiration(String value) {
+    public PublishDslBuilder expiration(String value) {
         this.wrapped = wrapped.expiration(toStringExpression(value));
         return this;
     }
 
-    public RequestReplyDslBuilder timestamp(String value) {
+    public PublishDslBuilder timestamp(String value) {
         this.wrapped = wrapped.timestamp(toExpression(value, java.util.Date.class));
         return this;
     }
 
-    public RequestReplyDslBuilder amqpType(String value) {
+    public PublishDslBuilder amqpType(String value) {
         this.wrapped = wrapped.amqpType(toStringExpression(value));
         return this;
     }
 
-    public RequestReplyDslBuilder userId(String value) {
+    public PublishDslBuilder userId(String value) {
         this.wrapped = wrapped.userId(toStringExpression(value));
         return this;
     }
 
-    public RequestReplyDslBuilder appId(String value) {
+    public PublishDslBuilder appId(String value) {
         this.wrapped = wrapped.appId(toStringExpression(value));
         return this;
     }
 
-    public RequestReplyDslBuilder clusterId(String value) {
+    public PublishDslBuilder clusterId(String value) {
         this.wrapped = wrapped.clusterId(toStringExpression(value));
         return this;
     }
 
-    public RequestReplyDslBuilder header(String key, String value) {
+    public PublishDslBuilder header(String key, String value) {
         this.wrapped = wrapped.header(key, toStringExpression(value));
         return this;
     }
 
-    public RequestReplyDslBuilder headers(Map<String, String> values) {
+    public PublishDslBuilder headers(Map<String, String> values) {
         var headersList = values
                 .entrySet()
                 .stream()
@@ -93,13 +90,7 @@ public class RequestReplyDslBuilder implements ActionBuilder {
         return this;
     }
 
-    public RequestReplyDslBuilder check(io.gatling.core.check.Check<AmqpProtocolMessage>... checks) {
-        this.wrapped = wrapped.check(scala.jdk.javaapi.CollectionConverters.asScala(Arrays.stream(checks).toList()).toSeq());
-        return this;
-    }
-
     public io.gatling.core.action.builder.ActionBuilder asScala(){
         return wrapped.build();
     }
-
 }
