+++ src/main/java/ru/tinkoff/gatling/javaapi/request/RequestReplyDslBuilder.java
@@ -1,85 +1,88 @@
-package ru.tinkoff.gatling.javaapi;
-
-import java.util.Map;
-import static io.gatling.javaapi.core.internal.Expressions.*;
+package ru.tinkoff.gatling.javaapi.request;
 
 import io.gatling.javaapi.core.ActionBuilder;
+import ru.tinkoff.gatling.amqp.request.AmqpProtocolMessage;
 import scala.Tuple2;
 
-public class PublishDslBuilder implements ActionBuilder {
+import java.util.Arrays;
+import java.util.Map;
 
-    private ru.tinkoff.gatling.amqp.request.PublishDslBuilder wrapped;
+import static io.gatling.javaapi.core.internal.Expressions.*;
+import static io.gatling.javaapi.core.internal.Expressions.toStringExpression;
 
-    public PublishDslBuilder(ru.tinkoff.gatling.amqp.request.PublishDslBuilder wrapped) {
+public class RequestReplyDslBuilder implements ActionBuilder {
+    private ru.tinkoff.gatling.amqp.request.RequestReplyDslBuilder wrapped;
+
+    public RequestReplyDslBuilder(ru.tinkoff.gatling.amqp.request.RequestReplyDslBuilder wrapped) {
         this.wrapped = wrapped;
     }
 
-    public PublishDslBuilder messageId(String value) {
+    public RequestReplyDslBuilder messageId(String value) {
         this.wrapped = wrapped.messageId(toStringExpression(value));
         return this;
     }
 
-    public PublishDslBuilder priority(Integer value) {
+    public RequestReplyDslBuilder priority(Integer value) {
         this.wrapped = wrapped.priority(toIntExpression(value.toString()));
         return this;
     }
 
-    public PublishDslBuilder contentType(String value) {
+    public RequestReplyDslBuilder contentType(String value) {
         this.wrapped = wrapped.contentType(toStringExpression(value));
         return this;
     }
 
-    public PublishDslBuilder contentEncoding(String value) {
+    public RequestReplyDslBuilder contentEncoding(String value) {
         this.wrapped = wrapped.contentEncoding(toStringExpression(value));
         return this;
     }
 
-    public PublishDslBuilder correlationId(String value) {
+    public RequestReplyDslBuilder correlationId(String value) {
         this.wrapped = wrapped.correlationId(toStringExpression(value));
         return this;
     }
 
-    public PublishDslBuilder replyTo(String value) {
+    public RequestReplyDslBuilder replyTo(String value) {
         this.wrapped = wrapped.replyTo(toStringExpression(value));
         return this;
     }
 
-    public PublishDslBuilder expiration(String value) {
+    public RequestReplyDslBuilder expiration(String value) {
         this.wrapped = wrapped.expiration(toStringExpression(value));
         return this;
     }
 
-    public PublishDslBuilder timestamp(String value) {
+    public RequestReplyDslBuilder timestamp(String value) {
         this.wrapped = wrapped.timestamp(toExpression(value, java.util.Date.class));
         return this;
     }
 
-    public PublishDslBuilder amqpType(String value) {
+    public RequestReplyDslBuilder amqpType(String value) {
         this.wrapped = wrapped.amqpType(toStringExpression(value));
         return this;
     }
 
-    public PublishDslBuilder userId(String value) {
+    public RequestReplyDslBuilder userId(String value) {
         this.wrapped = wrapped.userId(toStringExpression(value));
         return this;
     }
 
-    public PublishDslBuilder appId(String value) {
+    public RequestReplyDslBuilder appId(String value) {
         this.wrapped = wrapped.appId(toStringExpression(value));
         return this;
     }
 
-    public PublishDslBuilder clusterId(String value) {
+    public RequestReplyDslBuilder clusterId(String value) {
         this.wrapped = wrapped.clusterId(toStringExpression(value));
         return this;
     }
 
-    public PublishDslBuilder header(String key, String value) {
+    public RequestReplyDslBuilder header(String key, String value) {
         this.wrapped = wrapped.header(key, toStringExpression(value));
         return this;
     }
 
-    public PublishDslBuilder headers(Map<String, String> values) {
+    public RequestReplyDslBuilder headers(Map<String, String> values) {
         var headersList = values
                 .entrySet()
                 .stream()
@@ -90,7 +93,13 @@ public class PublishDslBuilder implements ActionBuilder {
         return this;
     }
 
+    public RequestReplyDslBuilder check(io.gatling.core.check.Check<AmqpProtocolMessage>... checks) {
+        this.wrapped = wrapped.check(scala.jdk.javaapi.CollectionConverters.asScala(Arrays.stream(checks).toList()).toSeq());
+        return this;
+    }
+
     public io.gatling.core.action.builder.ActionBuilder asScala(){
         return wrapped.build();
     }
+
 }
